Moved FrontController to Helper\ namespace
